# server/openapi.yaml
openapi: 3.1.0
info:
  title: Research Optimizer API
  description: |
    Public-facing API for the Research Optimizer platform. Served by the main
    server process which handles HTTPS termination, authentication, and proxying
    to internal services (gateway, cms).

    All `/api` endpoints require authentication unless noted otherwise. Authentication
    is via session cookie (after OAuth login) or `X-API-Key` header.
  version: 1.0.0
servers:
  - url: https://localhost
    description: Local development
paths:
...
  /api/admin/users/{id}:
    parameters:
      - name: id
        in: path
        required: true
        schema:
          type: integer
    get:
      operationId: getUser
      summary: Get user by ID
      tags: [Admin]
      security:
        - sessionCookie: []
        - apiKey: []
      responses:
        "200":
          description: User details
    delete:
      operationId: deleteUser
      summary: Delete a user
      tags: [Admin]
      security:
        - sessionCookie: []
        - apiKey: []
      responses:
        "200":
          description: User deleted
  /api/admin/users/{id}/usage:
    parameters:
      - name: id
        in: path
        required: true
        schema:
          type: integer
    get:
      operationId: getUserUsage
      summary: Get usage history for a user
      tags: [Admin]
      security:
        - sessionCookie: []
        - apiKey: []
      responses:
        "200":
          description: Usage records
...
  /api/admin/profile:
    post:
      operationId: updateProfile
      summary: Update own profile
      description: Any authenticated user can update their own firstName and lastName.
      tags: [Admin]
      security:
        - sessionCookie: []
        - apiKey: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                firstName:
                  type: string
                lastName:
                  type: string
      responses:
        "200":
          description: Profile updated
...
  /api/admin/analytics:
    get:
      operationId: getAnalytics
      summary: Aggregated usage analytics
      tags: [Admin]
      security:
        - sessionCookie: []
        - apiKey: []
      parameters:
        - name: startDate
          in: query
          schema:
            type: string
            format: date
        - name: endDate
          in: query
          schema:
            type: string
            format: date
        - name: groupBy
          in: query
          schema:
            type: string
            enum: [hour, day, week, month, user, model]
            default: day
        - name: userId
          in: query
          schema:
            type: integer
        - name: search
          in: query
          schema:
            type: string
          description: Only for groupBy=user
        - name: role
          in: query
          schema:
            type: string
          description: Only for groupBy=user
        - name: status
          in: query
          schema:
            type: string
          description: Only for groupBy=user
        - name: sortBy
          in: query
          schema:
            type: string
        - name: sortOrder
          in: query
          schema:
            type: string
            enum: [ASC, DESC]
        - name: limit
          in: query
          schema:
            type: integer
        - name: offset
          in: query
          schema:
            type: integer
      responses:
        "200":
          description: Analytics data
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    type: array
                    items:
                      type: object
                  meta:
                    type: object
components:
  securitySchemes:
    sessionCookie:
      type: apiKey
      in: cookie
      name: connect.sid
      description: Session cookie set after OAuth login
    apiKey:
      type: apiKey
      in: header
      name: X-API-Key
      description: API key (rsk_... format) for programmatic access
...